extracted workoutDetailsFOrm
