+++ frontend/src/components/workout/components/WorkoutDetails.tsx
@@ -1,25 +1,14 @@
 import React, { useState } from 'react'
-import { useForm, Controller, type SubmitHandler } from 'react-hook-form'
+import { useForm, type SubmitHandler } from 'react-hook-form'
 import useFetchProgram from '../../../hooks/useFetchProgram'
 import type Program from '../../../types/Program'
 import { useParams, useNavigate } from 'react-router-dom'
 import { useAuth0 } from '@auth0/auth0-react'
 import { fetchProgram } from '../../../utils/fetchProgram'
-import { formatRestTime } from '../../../utils/formatRestTime'
 import Box from '@mui/material/Box'
-import {
-  Button,
-  Typography,
-  Checkbox,
-  TextField,
-  Grid,
-  useMediaQuery,
-  FormControlLabel,
-  InputAdornment,
-  CircularProgress
-} from '@mui/material'
-import useTheme from '@mui/material/styles/useTheme'
+import { Button, Typography, CircularProgress } from '@mui/material'
 import { Footer } from '../../common/components'
+import { WorkoutDetailsForm } from './'
 
 interface ProgramFormInputs {
   exercises: Array<{
@@ -28,7 +17,10 @@ interface ProgramFormInputs {
     weight: number
     completed: boolean
   }>
+  exerciseIndex: number
+  control: any
 }
+
 export const WorkoutDetails: React.FC = () => {
   const [program, setProgram] = useState<Program>()
   const { user, getAccessTokenSilently } = useAuth0()
@@ -36,13 +28,11 @@ export const WorkoutDetails: React.FC = () => {
     programId: string
   }>()
   useFetchProgram(programId, getAccessTokenSilently, fetchProgram, setProgram)
-  const { control, handleSubmit } = useForm<ProgramFormInputs>({
+  const { handleSubmit } = useForm<ProgramFormInputs>({
     mode: 'onBlur'
   })
   const navigate = useNavigate()
-  const theme = useTheme()
-  const isSmallScreen = useMediaQuery(theme.breakpoints.down('sm'))
-  console.log('program: ', program)
+
   const onSubmit: SubmitHandler<ProgramFormInputs> = async (
     data
   ): Promise<void> => {
@@ -116,259 +106,11 @@ export const WorkoutDetails: React.FC = () => {
           <form id="program-form" onSubmit={handleSubmit(onSubmit)}>
             {program?.exercises.map((exercise, exerciseIndex) => {
               return (
-                <Box
+                <WorkoutDetailsForm
                   key={`exercise-${exerciseIndex}`}
-                  sx={{
-                    border: '2px solid black',
-                    width: '90vw',
-                    height: 'auto',
-                    margin: '16px',
-                    padding: '8px',
-                    borderRadius: '16px',
-                    display: 'flex',
-                    flexDirection: 'column',
-                    alignItems: 'center'
-                  }}
-                >
-                  <Typography
-                    variant="h5"
-                    fontWeight="bold"
-                    textAlign="center"
-                    sx={{ marginBottom: '16px' }}
-                  >
-                    {exercise.exerciseName}
-                  </Typography>
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      flexDirection: 'column',
-                      justifyContent: 'space-around',
-                      alignItems: 'center'
-                    }}
-                  >
-                    <Box
-                      sx={{
-                        display: 'flex',
-                        justifyContent: 'center',
-                        height: '60px'
-                      }}
-                    >
-                      <Controller
-                        name={`exercises.${exerciseIndex}.weight`}
-                        control={control}
-                        defaultValue={exercise.weight}
-                        rules={{
-                          required: 'Weight is required',
-                          min: { value: 1, message: 'Minimum value is 1' }
-                        }}
-                        render={({
-                          field: { onChange, value, ref, onBlur },
-                          fieldState: { error }
-                        }) => (
-                          <TextField
-                            id={`weightInput-${exerciseIndex}`}
-                            label="weight"
-                            name="weight"
-                            type="number"
-                            size="small"
-                            value={value}
-                            onBlur={onBlur}
-                            onChange={(e) => {
-                              const val = parseFloat(e.target.value)
-                              if (!isNaN(val) && val !== 0) {
-                                onChange(val)
-                              } else {
-                                onChange('')
-                              }
-                            }}
-                            inputRef={ref}
-                            onKeyDown={(e) => {
-                              if (['e', '-', '+'].includes(e.key)) {
-                                e.preventDefault()
-                              }
-                            }}
-                            inputProps={{
-                              inputMode: 'decimal',
-                              pattern: '[0-9]*'
-                            }}
-                            InputProps={{
-                              endAdornment: (
-                                <InputAdornment position="end">
-                                  lbs
-                                </InputAdornment>
-                              )
-                            }}
-                            InputLabelProps={{
-                              sx: {
-                                fontSize: isSmallScreen ? '.95rem' : '1rem'
-                              }
-                            }}
-                            sx={{
-                              width: '120px'
-                            }}
-                            error={!(error == null)}
-                            helperText={error != null ? error.message : null}
-                          />
-                        )}
-                      />
-                    </Box>
-                    <Box
-                      sx={{
-                        display: 'flex',
-                        justifyContent: 'space-evenly',
-                        flexWrap: 'wrap'
-                      }}
-                    >
-                      {Array.from(
-                        { length: exercise.sets.length },
-                        (_, setIndex) => {
-                          return (
-                            <Box
-                              key={`set-${exerciseIndex}-${setIndex + 1}`}
-                              sx={{
-                                display: 'flex',
-                                margin: '8px',
-                                flexDirection: 'column',
-                                maxWidth: '80px',
-                                alignItems: 'space-evenly'
-                              }}
-                            >
-                              <Typography variant="body1" textAlign="center">
-                                Set: {setIndex + 1}
-                              </Typography>
-                              <Grid
-                                container
-                                spacing={2}
-                                justifyContent="center"
-                                alignItems="center"
-                                sx={{ paddingLeft: '8px' }}
-                              >
-                                <Grid
-                                  item
-                                  xs={12}
-                                  sx={{
-                                    display: 'flex',
-                                    justifyContent: 'center',
-                                    margin: '8px'
-                                  }}
-                                >
-                                  <Box
-                                    sx={{
-                                      display: 'flex',
-                                      justifyContent: 'center',
-                                      height: '60px'
-                                    }}
-                                  >
-                                    <Controller
-                                      name={`exercises.${exerciseIndex}.sets.${setIndex}.reps`}
-                                      control={control}
-                                      defaultValue={
-                                        exercise.sets[setIndex].reps
-                                      }
-                                      rules={{
-                                        required: 'Rep is required',
-                                        min: {
-                                          value: 1,
-                                          message: 'Minimum value is 1'
-                                        }
-                                      }}
-                                      render={({
-                                        field: { onChange, value, ref, onBlur },
-                                        fieldState: { error }
-                                      }) => (
-                                        <TextField
-                                          id={`repsInput-${exerciseIndex}-${
-                                            setIndex + 1
-                                          }`}
-                                          label="reps"
-                                          name="reps"
-                                          type="number"
-                                          size="small"
-                                          value={isNaN(value) ? '' : value}
-                                          inputRef={ref}
-                                          onBlur={onBlur}
-                                          onChange={(e) => {
-                                            const val = parseInt(e.target.value)
-
-                                            if (!isNaN(val) && val !== 0) {
-                                              onChange(val)
-                                            } else {
-                                              onChange('')
-                                            }
-                                          }}
-                                          onKeyDown={(e) => {
-                                            if (
-                                              ['e', '-', '+'].includes(e.key)
-                                            ) {
-                                              e.preventDefault()
-                                            }
-                                          }}
-                                          inputProps={{
-                                            inputMode: 'numeric',
-                                            pattern: '[0-9]*'
-                                          }}
-                                          InputLabelProps={{
-                                            sx: {
-                                              fontSize: isSmallScreen
-                                                ? '.95rem'
-                                                : '1rem',
-                                              textAlign: 'center'
-                                            }
-                                          }}
-                                          sx={{
-                                            width: '70px',
-                                            marginBottom: '8px'
-                                          }}
-                                          error={!(error == null)}
-                                          helperText={
-                                            error != null ? error.message : null
-                                          }
-                                        />
-                                      )}
-                                    />
-                                  </Box>
-                                </Grid>
-                              </Grid>
-                            </Box>
-                          )
-                        }
-                      )}
-                    </Box>
-                    <Typography
-                      variant="body1"
-                      textAlign="center"
-                      sx={{ marginTop: '8px' }}
-                    >
-                      Rest: {formatRestTime(exercise.rest)}
-                    </Typography>
-                    <Controller
-                      name={`exercises.${exerciseIndex}.completed`}
-                      control={control}
-                      defaultValue={false}
-                      render={({ field: { onChange, value, ref } }) => (
-                        <FormControlLabel
-                          control={
-                            <Checkbox
-                              id={`completed-checkbox-${exerciseIndex}`}
-                              checked={value || false}
-                              onChange={() => {
-                                if (!value) {
-                                  onChange(true)
-                                } else {
-                                  onChange(false)
-                                }
-                              }}
-                              style={{
-                                color: value ? 'green' : 'initial'
-                              }}
-                            />
-                          }
-                          label="Completed?"
-                        />
-                      )}
-                    />
-                  </Box>
-                </Box>
+                  exerciseIndex={exerciseIndex}
+                  exercise={exercise}
+                />
               )
             })}
