+++ frontend/src/components/workout/components/index.ts
@@ -2,5 +2,12 @@ import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
 import { WorkoutDetails } from './WorkoutDetails'
 import { WorkoutList } from './WorkoutList'
 import { WorkoutHistory } from './WorkoutHistory'
+import { WorkoutDetailsForm } from './WorkoutDetailsForm'
 
-export { WorkoutDetails, WorkoutHistoryTableRow, WorkoutList, WorkoutHistory }
+export {
+  WorkoutDetails,
+  WorkoutHistoryTableRow,
+  WorkoutList,
+  WorkoutHistory,
+  WorkoutDetailsForm
+}
