+++ frontend/src/types/Exercise.ts
@@ -4,7 +4,8 @@ interface Exercise {
   exerciseName: string
   sets: Set[]
   rest: string
-  weight?: number
+  weight: number
+  completed: boolean
 }
 
 export default Exercise
